Replace endTurn data logic by event-based onTurnEnd and onOwnTurnEnd
Incidentally, this lets us fix a bunch of minions with end-of-turn effects
which should have been set as end-of-own-turn effects.

This also lets us fix the frozen logic which was not checking whether the
character had attacked during the turn before removing the frozen tag. Not to
mention it was never removed on Heroes.