Have heroes specifically inherit Windfury from their Weapon
Add some tests for Doomhammer with this.
This was a tough decision. Weapon Windfury is not an aura, and the game does
set Windfury on the hero when Doomhammer comes into play.
The slot system is being somewhat phased out in favour of this cleaner (yet
more verbose) approach. Here we still are able to set hero.windfury = True even
though it also looks in the weapon to see if it's set.