Reimplement auras
Nth time's the charm.
The Game object still holds a list of auras, but auras are now better-behaved.
They will actually buff minions affected by them, and clear buffs when they get
destroyed. This is possible thanks to a new "onUpdate" event.
They will also keep a cache of minions they already buffed so that they do not
constantly remove and reapply themselves (that would break a lot of rules).