Get rid of inheritance on XMLCard
This commit removes all the inheritance weirdness that we had
going on.
XML is now parsed only once at initialization time. We build a
card database out of it and make it all importable in the new
fireplace.cards module (previously fireplace.carddata).
This also allows us to get rid of inheritance on Card in the
card definitions themselves. Incidentally, this lets us clean up
the imports in every card set and we now just do:
  from ..utils import *

Since we renamed carddata to cards, it's also an excellent
occasion to reorganize the card sets. Where we were previously
splitting between spells and minions, per class and very
inconsistently, we are now splitting between three core sets:
game, classic and naxxramas. The "removed" set is also there but
is not actually imported.

We are also updating hearthstone-data to support GVG. There was
a pretty significant change in the format which prompted this
whole rewrite.

Peace.