Check for has_target() when triggering the action
There is a lot that can happen during a Battlecry which causes a
target that was previously available not to be available anymore.
The other way around is also possible. For example, Rend Blackhand
considered itself as a valid target (until the previous commit),
so it would think a target is available when really it isn't.

This also avoids duplicating the has_target() logic for eg.
REQ_TARGET_IF_AVAILABLE_AND_DRAGON_IN_HAND...

Fixes #114