Implement Mulligan
Careful! This means a Mulligan phase is now forced on both players.
The players will *not* be able to end the turn, unless their mulligan
has been completed.

To "skip" a mulligan, one can simply do an empty mulligan with
`player.choice.choose()`. To remove the Mulligan phase altogether,
use a BaseGame instead of a Game object.