Fix Conceal and Finicky Cloakfield to behave like it does in-game
The cards only affect non-stealthed minions. They place a buff on
the minion which unstealthes it at the beginning of the next owner's
turn. It's prone to bugs due to possible overlap between cards such
as Master of Disguise, but it's also a lot easier to simulate than
a more proper implementation - so we go with that for now.

Various tests here:
https://www.youtube.com/watch?v=Tk1qeAm7Qjs