Revamp Choose Ones by making Card.choose_cards a list of playable cards
This is slightly closer to the upstream Hearthstone model which creates
separate entities for chosen cards from the beginning of the game.
In our case, we do it when the card arrives in the hand, that's more or
less the same thing.
This also gives us the ability to do deeper introspection based on the
chosen cards, such as getting their target list.
This also retains a Card.play(choose=id) API, although it is now just a
helper, internally calling Card.choose_cards.filter(id=id)[0].play(..).

Fixes #131