Create a CardDB class for fireplace.cards.db
This removes the fireplace.cards.<id> functionality, making
fireplace.cards.db the only point of access to the list of merged
cards.

Additionally, it moves the merge() and filter() functionality to the
CardDB class. fireplace.cards.filter() is now an alias to db.filter().

The merge is performed as lazily as possible. This is necessary because
the previous commit accesses members of fireplace.cards.<set> during
the bootstrap process, which would otherwise break as a merge would
be attempted with an empty CardDefs.xml file.