Play cards with a choose argument instead of reparenting
This cleans up the Choose One logic. Before, we would play the child
card and get its parent back in the Play action. This made the real
source of the battlecry very hard to track.
Battlecry sources did not matter until now as they were always the
same as the origin card.
For Djinni of Zephyrs, this is no longer true.