Implement a proper Mulligan phase, blocking game start until completed
The game no longer automatically continues after the Mulligan choices
have been offered. Instead, the first turn will begin as a callback
at the completion of both Mulligan choices.

Adds an extensive test for mulligan logic.

Note that cards are still not replaced in-place... if only some cards
are replaced, the hand will always have the initial cards on the left
and the replacement ones on the right.

Touches #329

Closes #334